Use a config file for settings and make update functions static
We remove the "set defaults" function in favour of the install
hook that represents the default state of the module. The default
settings are taken out to a configuration file.

The update hook is changed to always perform the same action. This
way if the defaults ever change then the update path behaviour will
not. This keeps update predictable and doesn't break them.